bug-fix: fix the infinite loop in ram_find_and_save_block()
In the case that last_seen_block is NULL and there is no dirty block, we
will loop infinitely along the ram_list.blocks list to find next dirty
block because we can't identify that we have iterated all the blocks.
Simply set last_seen_block to the first block if it is NULL, which will
fix this bug.

Signed-off-by: Le Tan <tamlokveer@gmail.com>