persist logic fix: donot put prefeted page to checkpoint
Persist logic fix: donot put prefeted page to checkpoint.
In smc_pml_persist_unprefetched_pages, if a dirty is not prefetched
in a round, we put it to checkpoint(set the bit in migration_bitmap)
but if this page is prefetched in another round later within the same
epoch, we should clear the bit.

Signed-off-by: ShiruRen <renshiru2000@gmail.com>